+++ src/hook_handlers/solidity.ts
@@ -1,4 +1,4 @@
-import { uploadSelectors } from '../logic.js';
+import { uploadSelectors } from '../lib/four_byte_uploader.js';
 import { isCi } from '@nomicfoundation/hardhat-utils/ci';
 import type { SolidityHooks } from 'hardhat/types/hooks';
